Reorganize file functions by what they do
