+++ src/lib/cache/index.ts
@@ -1,5 +1,5 @@
 export { default as cacheBuildTimestampExists } from './cache-build-timestamp-exists';
 export { default as cacheEntryExists } from './cache-entry-exists';
 export { default as getCache } from './get-cache';
-export { default as getLocalPackages } from './get-local-packages';
-export { default as getMatchingLocalPackages } from './get-matching-local-packages';
+export { default as getLocalPackages } from '../package/get-local-packages';
+export { default as getMatchingLocalPackages } from '../package/get-matching-local-packages';
