+++ src/lib/do-task.ts
@@ -1,7 +1,7 @@
 import { SpawnOptions } from 'child_process';
 import * as path from 'path';
 
-import getPackageDir from './get-package-dir';
+import getPackageDir from './package/get-package-dir';
 import runProcess from './run-process';
 
 export default function doTask(argv: any, taskName: string, packageName: string) {
