+++ src/lib/index.ts
@@ -2,7 +2,7 @@ import * as cache from './cache';
 export { cache };
 
 export { default as doTask } from './do-task';
-export { default as getMatchingPackageTasks } from './get-matching-package-tasks';
-export { default as getpackageDir } from './get-package-dir';
-export { default as getPackagePjson } from './get-package-pjson';
-export { default as getPackageTasks } from './get-package-tasks';
+export { default as getMatchingPackageTasks } from './package/get-matching-package-tasks';
+export { default as getpackageDir } from './package/get-package-dir';
+export { default as getPackagePjson } from './package/get-package-pjson';
+export { default as getPackageTasks } from './package/get-package-tasks';
