+++ src/lib/package/get-local-packages.ts
@@ -3,7 +3,7 @@ import * as path from 'path';
 import * as nodeGlob from 'glob';
 
 // This may eventually be configurable.
-import { PACKAGE_DIR } from './environment';
+import { PACKAGE_DIR } from '../environment';
 
 // This will be the packages on disk.
 let localPackages: string[];
