+++ src/lib/package/get-package-dir.ts
