+++ src/lib/package/get-package-pjson.ts
